fix(scheduled-tasks): address cron validation review
